[utils] split geometry utilities
